refactor: view 분리
